refactor: ridsl -> ril (richer-input language)
