+++ pkg/ril/dump.go
@@ -1,4 +1,4 @@
-package ridsl
+package ril
 
 import "fmt"
