+++ pkg/ril/example_test.go
@@ -1,59 +1,59 @@
-package ridsl_test
+package ril_test
 
 import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/ooni/2023-05-richer-input/pkg/ridsl"
+	"github.com/ooni/2023-05-richer-input/pkg/ril"
 	"github.com/ooni/probe-engine/pkg/runtimex"
 )
 
 func Example() {
 	// create a function describing a nettest workflow
-	function := ridsl.Compose(
+	function := ril.Compose(
 		// VoidType -> DomainNameType
-		ridsl.DomainName("www.example.com"),
+		ril.DomainName("www.example.com"),
 
 		// DomainNameType -> DNSLookupResultType
-		ridsl.DNSLookupParallel(
-			ridsl.DNSLookupGetaddrinfo(),
-			ridsl.DNSLookupUDP("8.8.4.4:53"),
+		ril.DNSLookupParallel(
+			ril.DNSLookupGetaddrinfo(),
+			ril.DNSLookupUDP("8.8.4.4:53"),
 		),
 
 		// DNSLookupResultType -> ListOfEndpointType
-		ridsl.MakeEndpointsForPort(443),
+		ril.MakeEndpointsForPort(443),
 
 		// ListOfEndpointType -> VoidType
-		ridsl.NewEndpointPipeline(
+		ril.NewEndpointPipeline(
 			// EndpointType -> TCPConnectionType
-			ridsl.TCPConnect(),
+			ril.TCPConnect(),
 
 			// TCPConnectionType -> TCPConnectionType
 			//
 			// Here we conditionally insert into the pipeline a passthrough filter
 			// that may not exist on the target OONI Probe instance. If it does not
-			// exist, the [riengine] will replace this with an identity Func.
-			ridsl.IfFuncExists(&ridsl.Func{
+			// exist, the [ric] will replace this with an identity Func.
+			ril.IfFuncExists(&ril.Func{
 				Name:       "foobar_check_tcp_connect",
-				InputType:  ridsl.TCPConnectionType,
-				OutputType: ridsl.TCPConnectionType,
+				InputType:  ril.TCPConnectionType,
+				OutputType: ril.TCPConnectionType,
 				Arguments:  nil,
-				Children:   []*ridsl.Func{},
+				Children:   []*ril.Func{},
 			}),
 
 			// TCPConnectionType -> TLSConnectionType
-			ridsl.TLSHandshake(),
+			ril.TLSHandshake(),
 
 			// TLSConnectionType -> HTTPRoundTripResponseType
-			ridsl.HTTPRoundTrip(),
+			ril.HTTPRoundTrip(),
 
 			// HTTPRoundTripResponseType -> VoidType
-			ridsl.HTTPReadResponseBodySnapshot(),
+			ril.HTTPReadResponseBodySnapshot(),
 		),
 	)
 
 	// convert the function to an ASTNode
-	node := ridsl.Compile(function)
+	node := ril.ExportASTNode(function)
 
 	// serialize to JSON
 	data := runtimex.Try1(json.Marshal(node))
