+++ pkg/ril/types_test.go
@@ -1,4 +1,4 @@
-package ridsl
+package ril
 
 import "testing"
